Add strategy pulling speech gaps to sentence start
The strategy pulls out gaps (speech breaks) as found in text. It
combines all the gaps into one duration and moves the break to the
border of the sentence, prepending sentence.

It is an experiment. Probably, the dub would feel closer to the
real person interpreting, with the original soundtrack voice starting
and translated dub kicking in a little later.