Add Azure as a text-to-speech provider (#49)
Refactor the way speech parameters are conveyed in SSML so that SSML is
compatible with Azure (strict XML namespace, speech rate goes to prosody
element and so on). Also refactor speech rate picking method so that
rate goes inside SSML and not as external parameter, as Azure does not
get the latter.

Add implementation of Azure cognitive services for text-to-speech.

Add two new voices: Bill and Melinda, implemented with Azure TTS. The
speech.py will decide which provider to use depending on the voice
requested. You can mix-and-match voices from Google and Azure.

Add requirement for two new env vars: AZURE_REGION and
AZURE_SUBSCRIPTION_KEY, please see updated Readme.