Make all characters speak all languages (#72)
* Make all characters speak all languages

As a result, whatever character is used in the transcript, it will get a
voice. Since, however, providers vary in count of voices available, the
voices are sometimes re-used. For example, Alonzo Church and Alan Turing
might speak differently in English but would speak same voice in
Ukrainian.

Fix #58, in a way.