Improve error message handling (#67)
Improve error message handling

Now the error can get from a deep service to a user interface like chat.
At the same time, code is not relying on 500 errors and default handling
now. User input, configuration mismatch and similar errors are now
passed as 4XX HTTP with details in the body.

Remove a lot of explicit error handling. Pass the error via HTTP layer
in a default way so that we do not need to reconstruct it on the caller
level in a special way.